MODDICORE-63:Support matching by setting name in case schema contains UUID 1. Code improvements.
