+++ src/main/java/org/folio/processing/matching/matcher/Matcher.java
@@ -2,12 +2,9 @@ package org.folio.processing.matching.matcher;
 
 import io.vertx.core.json.JsonObject;
 
-import org.apache.commons.lang.StringUtils;
 import org.folio.DataImportEventPayload;
 import org.folio.MatchDetail;
 import org.folio.MatchProfile;
-import org.folio.processing.matching.MatchingManager;
-import org.folio.processing.matching.loader.LoadResult;
 import org.folio.processing.matching.loader.MatchValueLoader;
 import org.folio.processing.matching.loader.query.LoadQuery;
 import org.folio.processing.matching.loader.query.LoadQueryBuilder;
@@ -32,11 +29,6 @@ public interface Matcher {
     // Only one matching detail is expected in first implementation,
     // in future matching will support multiple matching details combined in logic expressions
     MatchDetail matchDetail = matchProfile.getMatchDetails().get(0);
-    if (matchDetail.getIncomingMatchExpression().getStaticValueDetails() != null && StringUtils.isNotEmpty(matchDetail.getIncomingMatchExpression().getStaticValueDetails().getText())) {
-      matchDetail.getIncomingMatchExpression().getStaticValueDetails()
-        .setText(MatchingManager.retrieveIdFromContext(matchDetail, eventPayload));
-    }
-    eventPayload.getContext().remove("MATCHING_PARAMETERS_RELATIONS");
 
     Value value = matchValueReader.read(eventPayload, matchDetail);
     LoadQuery query = LoadQueryBuilder.build(value, matchDetail);
