+++ src/main/java/org/folio/processing/matching/reader/StaticValueReaderImpl.java
@@ -14,7 +14,14 @@ import org.folio.rest.jaxrs.model.StaticValueDetails;
 import static java.util.Objects.nonNull;
 import static org.folio.rest.jaxrs.model.MatchExpression.DataValueType.STATIC_VALUE;
 
+import io.netty.util.internal.StringUtil;
+import io.vertx.core.json.JsonArray;
+import io.vertx.core.json.JsonObject;
+
 public class StaticValueReaderImpl implements MatchValueReader {
+
+  private static final String MAPPING_PARAMS = "MAPPING_PARAMS";
+  private static final String RELATIONS = "MATCHING_PARAMETERS_RELATIONS";
 
   @Override
   public Value read(DataImportEventPayload eventPayload, MatchDetail matchDetail) {
@@ -22,8 +29,8 @@ public class StaticValueReaderImpl implements MatchValueReader {
     if (matchExpression.getDataValueType() == STATIC_VALUE && nonNull(matchExpression.getStaticValueDetails())) {
       StaticValueDetails staticValueDetails = matchExpression.getStaticValueDetails();
       switch (staticValueDetails.getStaticValueType()) {
-        case TEXT: return obtainStringValue(staticValueDetails.getText());
-        case NUMBER: return obtainStringValue(staticValueDetails.getNumber());
+        case TEXT: return obtainStringValue(matchDetail, eventPayload);
+        case NUMBER: return obtainNumberValue(staticValueDetails.getNumber());
         case EXACT_DATE: return obtainDateValue(staticValueDetails.getExactDate(), staticValueDetails.getExactDate());
         case DATE_RANGE: return obtainDateValue(staticValueDetails.getFromDate(), staticValueDetails.getToDate());
         default: return MissingValue.getInstance();
@@ -37,11 +44,34 @@ public class StaticValueReaderImpl implements MatchValueReader {
     return incomingRecordType == EntityType.STATIC_VALUE;
   }
 
-  private Value obtainStringValue(String value) {
+  private Value obtainStringValue(MatchDetail matchDetail, DataImportEventPayload eventPayload) {
+    String id = retrieveIdFromContext(matchDetail, eventPayload);
+    return nonNull(id) ? StringValue.of(id) : MissingValue.getInstance();
+  }
+
+  private Value obtainNumberValue(String value) {
     return nonNull(value) ? StringValue.of(value) : MissingValue.getInstance();
   }
 
   private Value obtainDateValue(Date from, Date to) {
     return nonNull(from) && nonNull(to) ? DateValue.of(from, to) : MissingValue.getInstance();
+  }
+
+  public static String retrieveIdFromContext(MatchDetail matchDetail, DataImportEventPayload eventPayload) {
+    JsonObject matchingParams = new JsonObject(eventPayload.getContext().get(MAPPING_PARAMS));
+    JsonObject relations = new JsonObject(eventPayload.getContext().get(RELATIONS));
+    String relation = String.valueOf(relations.getJsonObject("matchingRelations")
+      .getMap().get(matchDetail.getExistingMatchExpression().getFields().get(0).getValue()));
+    JsonArray jsonArray = matchingParams.getJsonArray(relation);
+
+    for (int i = 0; i < jsonArray.size(); i++) {
+      if (jsonArray.getJsonObject(i).getString("name")
+        .equals(matchDetail.getIncomingMatchExpression().getStaticValueDetails().getText().trim())) {
+        JsonObject result = jsonArray.getJsonObject(i);
+        return result.getString("id");
+      }
+    }
+    eventPayload.getContext().remove(RELATIONS);
+    return StringUtil.EMPTY_STRING;
   }
 }
